Fix readme and comments
